Added a generic Renderer module.
This module and associated struct allows a rendering pipeline to be defined in
terms of, at the moment, simplistic steps. Each step has an associated Shader,
a number of textures (some which may be shared using Rc-s) and static VBOs.

This is a modest start of the 'unified rendering model' that I'm trying to
implement. Importantly, in the future we'll need dynamic/stream VBOs and
dynamic index buffers into static VBOs.