Implemented shader lighting effects.
This implements all the lighting effects present in Doom 1 & 2 (flash, strobe,
glow, with sector-level syncronisation). The implementation is completely done
in the shaders, keeping the VBOs static.

This commit also fixes a problem which crashed the renderer on E4M7 due to
degenerate polygons after the simplification procedure; this polys are now
simply dropped.