WadName type instead of [u8] and Vec-s.
Painful yet important commit. Replaced the WadName type synonym (which required
the use of Vec<u8> and &[u8] occasionally and did not support Eq & Hashing) to
a real type which supports all that as well as painless str, String and &[u8]
conversions.

Added ```--load-all``` command line option which simply loads all levels and
quits. Useful for diagnostics, ensuring no level loading fails.

Fixed a typo in one of the texture names which prevented it to load correctly.
Spotted thanks to the new WadName class.