Added TOML wad metadata.
A lot of stuff was hardcoded in the original doom; instead of replicating that,
I'm moving all asset related data not stored in the WAD file into a separate
TOML file.

For now this only includes the sky to level mapping, but it will be where I
store thing type -> sprite mappings as well, for instance.