Reverted to GLSL 330 on non-linux OS-s.
This is a quick fix, it isn't really the way we should do this. The commit adds
the ShaderLoader struct which deals with loading shaders from files and
prepends the version directive based on the platform.

In future, ShaderLoader should be the one dealing with preprocessing shader
files in general (#include, #ifdef etc.)